Fixed syntax errors; Improved the batching object
Cleaned up a lot of errors in the previews commit - so please next time validate the YAML before committing it. 

We kept most of the changes like the /scopes/ but placed the trackerId in the body JSON rather then repeating it in each and every record in the array. so now the JSON looks like:
{
 trackerId: 'bla',
 events: [{}, {}]
}